requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.2
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
boto3==1.35.0
python-multipart==0.0.9
dodopayments==1.66.0
playwright==1.49.0
greenlet==3.1.1
resend==2.0.0
jinja2==3.1.4
firebase-admin==6.5.0
phonenumbers==8.13.0
